Project6
I made a few tweaks to my previous activities, and I programmed and wired up project 6. It can count item going on a conveyor belt using a beam of light. It will warn the user if there's a blockage, and it will automatically empty the collector when 30 items are counted.